Moving tests back to src/test.
